Minor refactors - Memoize in createcharacter - use functors in memory functions - move char functions to common char module
