+++ common/memory.ts
@@ -3,6 +3,8 @@ import { defaultPresets } from './presets'
 import { BOT_REPLACE, SELF_REPLACE } from './prompt'
 import { Encoder } from './tokenize'
 
+export const BUNDLED_CHARACTER_BOOK_ID = '__bundled__characterbook__'
+
 export type MemoryOpts = {
   user: AppSchema.User
   chat: AppSchema.Chat
@@ -25,13 +27,43 @@ type Match = {
 
   /** The position within the prompt that it was found. Larger means closer to the end */
   index: number
-
   entry: AppSchema.MemoryEntry
-
   tokens: number
-
   text: string
 }
+
+export type CharacterBook = {
+  name?: string
+  description?: string
+  scan_depth?: number /** agnai: "Memory: Chat History Depth" */
+  token_budget?: number /** agnai: "Memory: Context Limit" */
+  recursive_scanning?: boolean /** no agnai equivalent. whether entry content can trigger other entries */
+  extensions: Record<string, any>
+  entries: Array<{
+    keys: Array<string>
+    content: string
+    extensions: Record<string, any>
+    enabled: boolean
+    insertion_order: number /** if two entries inserted, lower "insertion order" = inserted higher */
+    case_sensitive?: boolean
+
+    // FIELDS WITH NO CURRENT EQUIVALENT IN SILLY
+    name?: string /** not used in prompt engineering */
+    priority?: number /** if token budget reached, lower priority value = discarded first */
+
+    // FIELDS WITH NO CURRENT EQUIVALENT IN AGNAI
+    id?: number /** not used in prompt engineering */
+    comment?: string /** not used in prompt engineering */
+    selective?: boolean /** if `true`, require a key from both `keys` and `secondary_keys` to trigger the entry */
+    secondary_keys?: Array<string> /** see field `selective`. ignored if selective == false */
+    constant?: boolean /** if true, always inserted in the prompt (within budget limit) */
+    position?:
+      | 'before_char'
+      | 'after_char' /** whether the entry is placed before or after the character defs */
+  }>
+}
+
+type CharacterBookEntry = CharacterBook['entries'][number]
 
 /**
  * When determine the insertion order of an entry:
@@ -123,42 +155,58 @@ function prep(str: string, safe?: boolean) {
   return next.toLowerCase().replace(/\-/g, '\\-')
 }
 
-export const BUNDLED_CHARACTER_BOOK_ID = 'characterBook3872489385'
-
-export const characterBookToAgnaiMemory = (cb: CharacterBook): AppSchema.MemoryBook => ({
-  kind: 'memory',
-  _id: BUNDLED_CHARACTER_BOOK_ID,
-  name: cb.name ?? 'Character book',
-  description: cb.description ?? '',
-  userId: 'characterBook',
-  entries: cb.entries.map((cbEntry) => ({
-    name: cbEntry.name ?? 'Unnamed',
-    entry: cbEntry.content,
-    keywords: cbEntry.keys,
-    priority: cbEntry.priority ?? 100,
-    weight: cbEntry.insertion_order ?? 100,
-    enabled: cbEntry.enabled,
+export function characterBookToNative(cb: CharacterBook): AppSchema.MemoryBook {
+  return {
+    kind: 'memory',
+    _id: BUNDLED_CHARACTER_BOOK_ID,
+    name: cb.name ?? 'Character book',
+    description: cb.description ?? '',
+    userId: 'characterBook',
+    entries: cb.entries.map(memoryEntryToNative),
 
     // currently unsupported V2 fields which are here so that we don't destroy them
-    id: cbEntry.id,
-    comment: cbEntry.comment,
-    selective: cbEntry.selective,
-    secondaryKeys: cbEntry.secondary_keys,
-    constant: cbEntry.constant,
-    position: cbEntry.position,
-  })),
+    scanDepth: cb.scan_depth,
+    tokenBudget: cb.token_budget,
+    recursiveScanning: cb.recursive_scanning,
+    extensions: cb.extensions,
+  }
+}
 
-  // currently unsupported V2 fields which are here so that we don't destroy them
-  scanDepth: cb.scan_depth,
-  tokenBudget: cb.token_budget,
-  recursiveScanning: cb.recursive_scanning,
-  extensions: cb.extensions,
-})
+export function nativeToCharacterBook(memory: AppSchema.MemoryBook): CharacterBook {
+  return {
+    name: memory.name,
+    description: memory.description,
+    entries: memory.entries.map(nativeToMemoryEntry),
 
-export const agnaiMemoryToCharacterBook = (memory: AppSchema.MemoryBook): CharacterBook => ({
-  name: memory.name,
-  description: memory.description,
-  entries: memory.entries.map((memoryEntry) => ({
+    // currently unsupported V2 fields which are here so that we don't destroy them
+    extensions: memory.extensions ?? {},
+    scan_depth: memory.scanDepth,
+    token_budget: memory.tokenBudget,
+    recursive_scanning: memory.recursiveScanning,
+  }
+}
+
+function memoryEntryToNative(entry: CharacterBookEntry): AppSchema.MemoryEntry {
+  return {
+    name: entry.name ?? 'Unnamed',
+    entry: entry.content,
+    keywords: entry.keys,
+    priority: entry.priority ?? 100,
+    weight: entry.insertion_order ?? 100,
+    enabled: entry.enabled,
+
+    // currently unsupported V2 fields which are here so that we don't destroy them
+    id: entry.id,
+    comment: entry.comment,
+    selective: entry.selective,
+    secondaryKeys: entry.secondary_keys,
+    constant: entry.constant,
+    position: entry.position,
+  }
+}
+
+function nativeToMemoryEntry(memoryEntry: AppSchema.MemoryEntry): CharacterBookEntry {
+  return {
     keys: memoryEntry.keywords,
     content: memoryEntry.entry,
     extensions: {},
@@ -174,40 +222,5 @@ export const agnaiMemoryToCharacterBook = (memory: AppSchema.MemoryBook): Charac
     secondary_keys: memoryEntry.secondaryKeys,
     constant: memoryEntry.constant,
     position: memoryEntry.position,
-  })),
-
-  // currently unsupported V2 fields which are here so that we don't destroy them
-  extensions: memory.extensions ?? {},
-  scan_depth: memory.scanDepth,
-  token_budget: memory.tokenBudget,
-  recursive_scanning: memory.recursiveScanning,
-})
-
-export type CharacterBook = {
-  name?: string
-  description?: string
-  scan_depth?: number // agnai: "Memory: Chat History Depth"
-  token_budget?: number // agnai: "Memory: Context Limit"
-  recursive_scanning?: boolean // no agnai equivalent. whether entry content can trigger other entries
-  extensions: Record<string, any>
-  entries: Array<{
-    keys: Array<string>
-    content: string
-    extensions: Record<string, any>
-    enabled: boolean
-    insertion_order: number // if two entries inserted, lower "insertion order" = inserted higher
-    case_sensitive?: boolean
-
-    // FIELDS WITH NO CURRENT EQUIVALENT IN SILLY
-    name?: string // not used in prompt engineering
-    priority?: number // if token budget reached, lower priority value = discarded first
-
-    // FIELDS WITH NO CURRENT EQUIVALENT IN AGNAI
-    id?: number // not used in prompt engineering
-    comment?: string // not used in prompt engineering
-    selective?: boolean // if `true`, require a key from both `keys` and `secondary_keys` to trigger the entry
-    secondary_keys?: Array<string> // see field `selective`. ignored if selective == false
-    constant?: boolean // if true, always inserted in the prompt (within budget limit)
-    position?: 'before_char' | 'after_char' // whether the entry is placed before or after the character defs
-  }>
+  }
 }
