+++ web/shared/TextInput.tsx
@@ -76,7 +76,7 @@ const TextInput: Component<{
             value={value()}
             class={'form-field focusable-field w-full rounded-xl px-4 py-2 ' + props.class}
             onkeyup={(ev) => props.onKeyUp?.(ev)}
-            onchange={(ev) => props.onChange?.(ev)}
+            onChange={(ev) => props.onChange?.(ev)}
             disabled={props.disabled}
             pattern={props.pattern}
           />
